% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_jsd}
\alias{calculate_jsd}
\title{Calculate Jensen-Shannon divergence}
\usage{
calculate_jsd(x1, x2, nbins = 50)
}
\arguments{
\item{x1}{Numeric vector (first distribution)}

\item{x2}{Numeric vector (second distribution)}

\item{nbins}{Number of bins for histogram (default: 50)}
}
\value{
Jensen-Shannon divergence
}
\description{
Calculate Jensen-Shannon divergence
}
\keyword{internal}
