% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitss.R
\name{combine_mi_results}
\alias{combine_mi_results}
\title{Combine multiple imputation results using Rubin's rules}
\usage{
combine_mi_results(gamma_m, V_m, alpha)
}
\arguments{
\item{gamma_m}{Vector of estimates from each imputation}

\item{V_m}{Vector of variances from each imputation}

\item{alpha}{Significance level}
}
\value{
List with combined estimate, standard error, and CI
}
\description{
Combine multiple imputation results using Rubin's rules
}
\keyword{internal}
