% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitss.R
\name{create_subclasses}
\alias{create_subclasses}
\title{Create subclasses based on propensity scores}
\usage{
create_subclasses(e_hat, W, n_subclasses, min_per_subclass)
}
\arguments{
\item{e_hat}{Estimated propensity scores}

\item{W}{Treatment assignment}

\item{n_subclasses}{Maximum number of subclasses}

\item{min_per_subclass}{Minimum units per treatment group per subclass}
}
\value{
List with subclass assignments and actual number of subclasses
}
\description{
Create subclasses based on propensity scores
}
\keyword{internal}
