% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitss.R
\name{estimate_propensity_score}
\alias{estimate_propensity_score}
\title{Estimate propensity scores}
\usage{
estimate_propensity_score(W, X)
}
\arguments{
\item{W}{Binary treatment vector}

\item{X}{Covariate matrix}
}
\value{
Vector of estimated propensity scores
}
\description{
Estimate propensity scores
}
\keyword{internal}
