% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitss.R
\name{fit_spline_model}
\alias{fit_spline_model}
\title{Fit spline model with orthogonalized covariates}
\usage{
fit_spline_model(Y, logit_e, X_ort, subclasses, n_subclasses)
}
\arguments{
\item{Y}{Outcome values}

\item{logit_e}{Logit of propensity scores}

\item{X_ort}{Orthogonalized covariates}

\item{subclasses}{Subclass assignments}

\item{n_subclasses}{Number of subclasses}
}
\value{
Fitted model object
}
\description{
Fit spline model with orthogonalized covariates
}
\keyword{internal}
