% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitss.R
\name{orthogonalize_covariates}
\alias{orthogonalize_covariates}
\title{Orthogonalize covariates with respect to propensity score}
\usage{
orthogonalize_covariates(X, e_hat)
}
\arguments{
\item{X}{Covariate matrix}

\item{e_hat}{Estimated propensity scores}
}
\value{
Orthogonalized covariate matrix
}
\description{
Orthogonalize covariates with respect to propensity score
}
\keyword{internal}
