% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitss.R
\name{perform_single_imputation}
\alias{perform_single_imputation}
\title{Perform single imputation}
\usage{
perform_single_imputation(data_list, n_subclasses)
}
\arguments{
\item{data_list}{List containing Y, W, X, X_ort, e_hat, subclasses}

\item{n_subclasses}{Number of subclasses}
}
\value{
List with imputed values and estimate
}
\description{
Perform single imputation
}
\keyword{internal}
