% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitss.R
\name{predict_from_spline}
\alias{predict_from_spline}
\title{Predict from fitted spline model}
\usage{
predict_from_spline(fit, logit_e_new, X_ort_new, subclasses_new)
}
\arguments{
\item{fit}{Fitted model object}

\item{logit_e_new}{Logit propensity scores for prediction}

\item{X_ort_new}{Orthogonalized covariates for prediction}

\item{subclasses_new}{Subclass assignments for prediction}
}
\value{
Predicted values with random draws from posterior
}
\description{
Predict from fitted spline model
}
\keyword{internal}
